feat(ShippingMethod.tsx): add dynamic behavior to ShippingMethod component based on the value of the "shippingAccount" field
The ShippingMethod component now uses the useFormContext hook from react-hook-form to access the form context and get the current value of the "shippingAccount" field. The component now dynamically sets the initial state of the "isPrepaid" variable based on whether the "shippingAccount" field has the value "prepaid". This allows the component to display different content based on the selected shipping method.