fix(PaymentMethodRadioGroup.tsx): set defaultValue of RadioGroup to getValues("paymentMethod") instead of hardcoding a value
The getValues function is added to the useFormContext hook to retrieve the current form values. This allows us to set the defaultValue of the RadioGroup component dynamically based on the value of the "paymentMethod" field in the form. This change improves the flexibility and maintainability of the code by removing the hardcoding of the defaultValue.