test(ShippingMethodRadioGroup): add test case for setting default value of shipping method
In the ShippingMethodRadioGroup test file, a new test case is added to ensure that the default value of the shipping method is set correctly. This is done by mocking the useFormContext hook and providing a mock implementation for the getValues function, which returns "DHL" as the default shipping method. Additionally, the register function is also mocked to prevent any errors related to unregistered fields.