feat(ui-dsc): add SearchPhenotypeList component to display a list of phenotypes
The SearchPhenotypeList component is added to the ui-dsc package. It is responsible for rendering a list of phenotypes. The component receives data, loadMore function, hasMore indicator, isLoadingMore indicator, and totalCount as properties. It uses Material-UI components such as Paper, List, CircularProgress, and Typography to create the UI. The component also utilizes the useIntersectionObserver hook from the dicty-hooks package to implement infinite scrolling functionality.