feat(ui-dsc): add SearchResultsHeader component to display search results header
The SearchResultsHeader component is added to the ui-dsc package. This component is responsible for displaying the header of the search results. It takes in two properties: `property` which represents the type of search results, and `description` which represents the item being searched for. The component renders a title with the `property` and "Search Results" appended to it, and a subtitle with the `description` wrapped in `<em>` tags.