refactor(SearchPhenotypeContainer.tsx): clean up code and improve error handling
The console.log statement for loading, error, and data has been removed as it is no longer needed. The code has been refactored to improve error handling. If there is an error (e) and data (d) is undefined, the GraphQLErrorPage component is rendered. This provides a better user experience by displaying the error message when there is an error and no data available.