feat(graphql_cache.ts): add support for ListStrainsWithPhenotypeQuery and ListStrainsWithAnnotationQuery
The graphql_cache.ts file has been modified to import the ListStrainsWithPhenotypeQuery and ListStrainsWithAnnotationQuery types from the dicty-graphql-schema package. This allows the code to work with these new query types. Additionally, new types and functions have been added to handle pagination and merging of the ListStrainsWithAnnotationQuery results. This change improves the functionality of the code by adding support for querying strains with annotations.