refactor(SearchPhenotypeContainer.tsx): remove unnecessary console.log statement and simplify match pattern
The console.log statement for loading, error, and data has been removed as it is no longer needed. The match pattern for handling the error has been simplified to only check for the presence of the error (e) and render the GraphQLErrorPage component accordingly. This improves code readability and removes unnecessary code.