fix(useSearchWithRouter.tsx): change variable name from hasTag to isAcceptingInput for better semantics
The variable name hasTag has been changed to isAcceptingInput to improve the clarity and readability of the code. This change reflects the purpose of the variable more accurately, indicating whether the user text input is being accepted or not.
Every logical check for the state variable, as well as state updates have been inverted accordingly.

Closes #464