fix(useSearchWithRouter.tsx): fix issue where the most recent search term is not removed when it has not been committed
Previously, when a search field option of the searchbox is selected, it was added to the array of Autocomplete's values, whether or not a value had been entered in for the search term.
If a value had not been added (a user inputs text and presses enter), no chip would be created for that value, but the Autocomplete would still recognize it as being a selected value, preventing the
search term from being selected subsequently.

The `onChange` function now checks if the most recent value is "active" (acceptingInput === true) and removes it from the values if it hasn't been committed. This ensures that the previous, incomplete search term  is properly removed when a new option is selected from the dropdown.