refactor(useSearchWithRouter.tsx): remove unnecessary setIsAcceptingInput(true) call in the "input" case of the switch statement
The setIsAcceptingInput(true) call in the "input" case of the switch statement is unnecessary as it is already being set to true in the useEffect hook. The comment in the "default" case of the switch statement is updated to reflect the correct behavior, which is clearing inputs and creating a chip for the completed field.