feat(login.tsx): add login page component
This commit adds a new file `login.tsx` which contains the `SignIn`
component. This component uses the `useLogto` hook from the
`@logto/react` library to check if the user is authenticated. If the
user is authenticated, it navigates to the home page (`<Navigate to="/"
/>`), otherwise it navigates to the callback page (`<Navigate
to="/callback" />`). This component will be used for the login
functionality in the Stock Center application.