refactor(darkenOrLighten.ts): simplify the logic for darkening or lightening a color
The logic for darkening or lightening a color has been simplified by removing unnecessary if-else statements. Instead of checking if the lightness value is greater than, less than, or equal to 0.5, a newLightnessValue variable is calculated based on the current lightness value. If the lightness value is greater than or equal to 0.5, the new lightness value is calculated by subtracting 0.1. Otherwise, the new lightness value is calculated by adding 0.1. This simplification improves readability and reduces code duplication.