feat(callback.tsx): add callback page for handling sign-in callback
A new file `callback.tsx` has been added to the `pages` directory in the
`stock-center` app. This file contains a functional component `Callback`
that imports `useHandleSignInCallback` from the `@logto/react` package
and `LoadingDisplay` from the `@dictybase/ui-dsc` package. The
`Callback` component uses the `useHandleSignInCallback` hook to handle
the sign-in callback and displays a loading display if the `isLoading`
flag is true. This page is used to handle the sign-in callback
functionality in the application.