feat(ui-dsc): add NotFoundError component for displaying a UI when an item is not found
The NotFoundError component is added to handle the display of a UI when an item is not found. It includes an error message prop that can be passed to it, although it is currently not used within the component. The component utilizes React Router for navigation and Material-UI for styling. It renders a grid layout with an image, heading, and paragraphs to inform the user that the page they are looking for cannot be found. It also provides links to the homepage and an option to add a page to the current route.