chore(package.json): update package name to "navbar" and remove unnecessary properties
The package name is updated to "navbar" to make it consistent with the frontendx monorepo. Unnecessary properties such as "private", "files", "engines", "module", "size-limit", "peerDependencies", "lint-staged", and devDependencies are removed to simplify the package.json file. The package version is also updated to "0.0.0" we no longer need to track its version.