chore(frontpage-components): update devDependencies in package.json
The devDependencies in the package.json files of frontpage-components
and hook-dsc packages have been updated. In frontpage-components, the
"@types/react" and "@types/react-dom" dependencies have been moved to
the top of the devDependencies section for better organization.
Additionally, the "eslint-config-custom" dependency has been added. In
hook-dsc, the "eslint-config-custom" dependency has been added.