feat(useSearchWithRouter.tsx): add renderOption function to customize the rendering of options in the autocomplete dropdown
The div element that was used to render the last value in the autocomplete dropdown has been replaced with the Chip component. This change improves the visual appearance and consistency with the rest of the options. Additionally, a new renderOption function has been added to customize the rendering of options in the autocomplete dropdown, allowing for more flexibility in how the options are displayed.