refactor(routes.tsx): refactor routeObject function to separate routes with and without authentication
The routeObject function has been refactored to separate routes that
require authentication from routes that do not. Routes without
authentication are now filtered and mapped to create the routeChildren
array. Routes with authentication are also filtered and mapped to create
the routeChildrenWithAuth array. The Protected component is then added
as a route element, with the routeChildrenWithAuth array as its
children. This allows for the support of protected routes in the
application.