feat(auth): add Callback component to handle sign-in callback
The Callback component is added to handle the sign-in callback. It uses
the `useHandleSignInCallback` hook from the `@logto/react` package to
handle the sign-in callback and redirects the user to the home page
("/") using the `useNavigate` hook from the `react-router-dom` package.
If the sign-in process is still loading, a loading display with 5 rows
is shown.