feat(HeaderWithAuth.tsx): add HeaderWithAuth component to handle authentication in the header
The HeaderWithAuth component is added to handle authentication in the
header. It imports necessary dependencies such as Header from
"@dictybase/header", match and P from "ts-pattern", useLogto and
UserInfoResponse from "@logto/react", useState and useEffect from
"react", and Router from "@remix-run/router". It also imports
LoginButton and LogoutButton components, as well as callbackPath and
homePath constants. The component takes in a clientRouter prop and uses
the useLogto hook to handle authentication logic. It conditionally
renders the Header component based on the authentication status and user
information. If the user is authenticated, it renders the LogoutButton
component with the user information, otherwise it renders the
LoginButton component.