feat(ui-dsc): add FieldOption component to render a colored chip with a label
The FieldOption component is added to the catalog directory in the ui-dsc package. It imports the Chip component from the Material-UI library and uses the makeStyles and createStyles functions to define styles for the chip. The chip is rendered with a label and a color specified by the fieldOptionColor constant. This component can be used to display a colored chip with a label in the UI.