feat(routeManager.tsx): add routeManager file to handle dynamic routes and build merged routes
The routeManager.tsx file is added to the functional directory in the
stock-center app. This file handles the creation of dynamic routes based
on the provided route configuration. It exports functions to generate
public and protected routes, as well as a function to build the merged
routes. The dynamicRoutesProperties type is defined to represent the
configuration of the dynamic routes. The pathParts function is used to
transform the route path into a format compatible with react-router-dom.
The mapToRouteObject function is used to map the route configuration to
a RouteObject. The publicRoutes function filters the dynamic routes to
generate an array of public routes. The protectedRoutes function filters
the dynamic routes to generate an array of protected routes. The
buildMergedRoutes function combines the public and protected routes and
adds additional routes for authentication.