fix(functional.tsx): add key to address field grid item
The shipping and payment form components  were being updated the formState was changing.
Since the grid item components that the form fields were wrapped in did not receive a key prop, they were being rerendered, causing the browser's tab focus to be lost.
This commit adds the key prop to each grid item that wraps an address text field.

Closes #463