feat(user/show.tsx): add display of user roles
The import of UserInfoResponse from "@logto/react" is no longer used in
the file, so it has been removed to improve code cleanliness. The user's
roles are now displayed in the UI using the match function from
fp-ts/Array. If the user has no roles, a message indicating that will be
shown. Otherwise, the roles will be displayed.