feat(UserMenuItem.tsx): add UserMenuItem component
The UserMenuItem component is added to the auth package. It is a reusable component that represents a menu item for a user. It takes an icon, a label, and an onClick event handler as props. The component renders a Material-UI MenuItem with a ListItemIcon and a ListItemText. This component can be used to create a menu for user-related actions in the application.