feat(routeManager.tsx): add support for private routes with role-based access control
The changes introduce a new feature to the route manager that allows the
creation of private routes. These routes are only accessible to users
with specific roles, enhancing the security of the application. The
'usePermify' hook from the '@permify/react-role' package is used to
implement role-based access control. The 'Private' component is used to
wrap the routes that should be private. The 'roles' property is added to
the 'PageComponentData' type to specify the roles that have access to a
particular route.